Append sample summary to single collection exercise instead of creating new endpoint
