+++ ras_backstage/__init__.py
@@ -23,7 +23,6 @@ api = Api(title='Ras-Backstage', default='info', default_label="")
 collection_exercise_api = Namespace('collection-exercise',
                                     path='/backstage-api/v1/collection-exercise')
 party_api = Namespace('party', path='/backstage-api/v1/party')
-sample_api = Namespace('sample', path='/backstage-api/v1/sample')
 sign_in_api = Namespace('sign-in', path='/backstage-api/v1/sign-in')
 secure_messaging_api = Namespace('secure-messaging', path='/backstage-api/v1/secure-message')
 survey_api = Namespace('survey', path='/backstage-api/v1/survey')
@@ -31,7 +30,6 @@ collection_instrument_api = Namespace('survey', path='/backstage-api/v1/collecti
 
 api.add_namespace(collection_exercise_api)
 api.add_namespace(party_api)
-api.add_namespace(sample_api)
 api.add_namespace(sign_in_api)
 api.add_namespace(secure_messaging_api)
 api.add_namespace(survey_api)
@@ -42,7 +40,6 @@ import ras_backstage.error_handlers  # NOQA # pylint: disable=wrong-import-posit
 from ras_backstage.resources.info import Info  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.sign_in.sign_in import SignIn  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.party.get_party_details import PartyDetails  # NOQA # pylint: disable=wrong-import-position
-from ras_backstage.resources.sample.sample import Sample  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.secure_messaging.get_message_list import GetMessagesList  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.secure_messaging.get_message import GetMessage  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.secure_messaging.update_label import RemoveUnreadLabel  # NOQA # pylint: disable=wrong-import-position
