+++ test/resources/test_collection_exercise.py
@@ -7,21 +7,32 @@ import requests_mock
 from ras_backstage import app
 
 
+collection_exercise_id = 'e33daf0e-6a27-40cd-98dc-c6231f50e84a'
+collection_instrument_id = 'f732afbe-c710-4c95-a8d3-6644833195a7'
+sample_summary_id = '08c191b8-e8b8-4920-b8de-87f85e536463'
+survey_id = 'cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87'
+test_period = '000000'
+test_short_name = 'test'
+
+url_sample_summary = f'{app.config["RM_SAMPLE_SERVICE"]}samples/samplesummary/{sample_summary_id}'
+url_link_sample = f'{app.config["RM_COLLECTION_EXERCISE_SERVICE"]}' \
+                  f'collectionexercises/link/{collection_exercise_id}'
 url_get_survey_by_short_name = f'{app.config["RM_SURVEY_SERVICE"]}surveys/shortname/test'
 url_ces = f'{app.config["RM_COLLECTION_EXERCISE_SERVICE"]}' \
-          'collectionexercises/survey/cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87'
+          f'collectionexercises/survey/{survey_id}'
 url_ce = f'{app.config["RM_COLLECTION_EXERCISE_SERVICE"]}' \
-         f'collectionexercises/e33daf0e-6a27-40cd-98dc-c6231f50e84a'
+         f'collectionexercises/{collection_exercise_id}'
 url_get_collection_instrument = f'{app.config["RAS_COLLECTION_INSTRUMENT_SERVICE"]}' \
                                    f'collection-instrument-api/1.0.2/collectioninstrument'
+
 with open('test/test_data/collection_exercise/collection_exercise.json') as json_data:
     collection_exercise = json.load(json_data)
 
 
 def _build_search_string():
     classifiers = {
-        "SURVEY_ID": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87",
-        "COLLECTION_EXERCISE": "e33daf0e-6a27-40cd-98dc-c6231f50e84a"
+        "SURVEY_ID": survey_id,
+        "COLLECTION_EXERCISE": collection_exercise_id
     }
     return urlencode({'searchString': json.dumps(classifiers)})
 
@@ -31,7 +42,7 @@ class TestCollectionExercise(unittest.TestCase):
     def setUp(self):
         self.app = app.test_client()
         self.survey = {
-            "id": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87",
+            "id": survey_id,
             "longName": "Business Register and Employment Survey",
             "shortName": "BRES",
             "surveyRef": "221"
@@ -43,8 +54,8 @@ class TestCollectionExercise(unittest.TestCase):
             "scheduledExecutionDateTime": "2017-05-15T00:00:00Z"
           },
           {
-            "id": "e33daf0e-6a27-40cd-98dc-c6231f50e84a",
-            "exerciseRef": "000000",
+            "id": collection_exercise_id,
+            "exerciseRef": test_period,
             "scheduledExecutionDateTime": "2017-08-12T00:00:00Z"
           }
         ]
@@ -52,16 +63,28 @@ class TestCollectionExercise(unittest.TestCase):
             {
                 "classifiers": {
                     "COLLECTION_EXERCISE": [
-                        "e33daf0e-6a27-40cd-98dc-c6231f50e84a"
+                        collection_exercise_id,
                     ],
                     "RU_REF": [],
-                    "SURVEY_ID": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87"
+                    "SURVEY_ID": survey_id,
                 },
                 "file_name": "file",
-                "id": "f732afbe-c710-4c95-a8d3-6644833195a7",
-                "surveyId": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87"
+                "id": collection_instrument_id,
+                "surveyId": survey_id
             }
         ]
+        self.sample_summary = {
+            "sampleSummaryPK": 1,
+            "id": sample_summary_id,
+            "state": "INIT",
+            "ingestDateTime": "2017-11-06T14:02:24.203+0000"
+        }
+        self.linked_sample = {
+            "collectionExerciseId": collection_exercise_id,
+            "sampleSummaryIds": [
+                sample_summary_id,
+            ]
+        }
 
     @requests_mock.mock()
     def test_single_collection_exercise(self, mock_request):
@@ -69,25 +92,27 @@ class TestCollectionExercise(unittest.TestCase):
         mock_request.get(url_get_survey_by_short_name, json=self.survey)
         mock_request.get(url_ces, json=self.collection_exercises)
         mock_request.get(url_ce, json=collection_exercise)
+        mock_request.get(url_link_sample, json=[sample_summary_id])
+        mock_request.get(url_sample_summary, json=self.sample_summary)
         search_string = _build_search_string()
         mock_request.get(f'{url_get_collection_instrument}?{search_string}', json=self.collection_instruments,
                          complete_qs=True)
 
         # When
-        response = self.app.get("/backstage-api/v1/collection-exercise/test/000000")
+        response = self.app.get(f'/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}')
         response_data = json.loads(response.data)
 
         # Then
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response_data['survey']['longName'],
                          'Business Register and Employment Survey')
-        self.assertEqual(response_data['collection_exercise']['name'], '000000')
+        self.assertEqual(response_data['collection_exercise']['name'], test_period)
 
     @requests_mock.mock()
     def test_single_collection_exercise_survey_fail(self, mock_request):
         mock_request.get(url_get_survey_by_short_name, status_code=500)
 
-        response = self.app.get("/backstage-api/v1/collection-exercise/test/000000")
+        response = self.app.get(f"/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}")
         response_data = json.loads(response.data)
 
         self.assertEqual(response.status_code, 500)
@@ -98,7 +123,7 @@ class TestCollectionExercise(unittest.TestCase):
         mock_request.get(url_get_survey_by_short_name, json=self.survey)
         mock_request.get(url_ces, status_code=500)
 
-        response = self.app.get("/backstage-api/v1/collection-exercise/test/000000")
+        response = self.app.get(f"/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}")
         response_data = json.loads(response.data)
 
         self.assertEqual(response.status_code, 500)
@@ -109,7 +134,7 @@ class TestCollectionExercise(unittest.TestCase):
         mock_request.get(url_get_survey_by_short_name, json=self.survey)
         mock_request.get(url_ces, json=[])
 
-        response = self.app.get("/backstage-api/v1/collection-exercise/test/000000")
+        response = self.app.get(f"/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}")
         response_data = json.loads(response.data)
 
         self.assertEqual(response.status_code, 404)
@@ -121,7 +146,7 @@ class TestCollectionExercise(unittest.TestCase):
         mock_request.get(url_ces, json=self.collection_exercises)
         mock_request.get(url_ce, status_code=500)
 
-        response = self.app.get("/backstage-api/v1/collection-exercise/test/000000")
+        response = self.app.get(f"/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}")
         response_data = json.loads(response.data)
 
         self.assertEqual(response.status_code, 500)
@@ -133,12 +158,14 @@ class TestCollectionExercise(unittest.TestCase):
         mock_request.get(url_get_survey_by_short_name, json=self.survey)
         mock_request.get(url_ces, json=self.collection_exercises)
         mock_request.get(url_ce, json=collection_exercise)
+        mock_request.get(url_link_sample, json=[sample_summary_id])
+        mock_request.get(url_sample_summary, json=self.sample_summary)
         search_string = _build_search_string()
         mock_request.get(f'{url_get_collection_instrument}?{search_string}', json=self.collection_instruments,
                          complete_qs=True)
 
         # When
-        response = self.app.get("/backstage-api/v1/collection-exercise/test/000000")
+        response = self.app.get(f"/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}")
         response_data = json.loads(response.data)
 
         # Then
@@ -156,7 +183,79 @@ class TestCollectionExercise(unittest.TestCase):
                          complete_qs=True, status_code=400)
 
         # When
-        response = self.app.get('/backstage-api/v1/collection-exercise/test/000000')
+        response = self.app.get(f'/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}')
+
+        # Then
+        self.assertEqual(response.status_code, 400)
+
+    @requests_mock.mock()
+    def test_get_sample_summary(self, mock_request):
+        # Given
+        mock_request.get(url_get_survey_by_short_name, json=self.survey)
+        mock_request.get(url_ces, json=self.collection_exercises)
+        mock_request.get(url_ce, json=collection_exercise)
+        mock_request.get(url_link_sample, json=[sample_summary_id])
+        mock_request.get(url_sample_summary, json=self.sample_summary)
+        search_string = _build_search_string()
+        mock_request.get(f'{url_get_collection_instrument}?{search_string}', json=self.collection_instruments,
+                         complete_qs=True)
+
+        # When
+        response = self.app.get(f"/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}")
+        response_data = json.loads(response.data)
+
+        # Then
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response_data['sample_summary'], self.sample_summary)
+
+    @requests_mock.mock()
+    def test_get_no_linked_summaries(self, mock_request):
+        # Given
+        mock_request.get(url_get_survey_by_short_name, json=self.survey)
+        mock_request.get(url_ces, json=self.collection_exercises)
+        mock_request.get(url_ce, json=collection_exercise)
+        search_string = _build_search_string()
+        mock_request.get(f'{url_get_collection_instrument}?{search_string}', json=self.collection_instruments,
+                         complete_qs=True)
+        mock_request.get(url_link_sample, status_code=204)
+
+        # When
+        response = self.app.get(f'/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}')
+
+        # Then
+        self.assertEqual(response.status_code, 204)
+
+    @requests_mock.mock()
+    def test_get_link_returns_error(self, mock_request):
+        # Given
+        mock_request.get(url_get_survey_by_short_name, json=self.survey)
+        mock_request.get(url_ces, json=self.collection_exercises)
+        mock_request.get(url_ce, json=collection_exercise)
+        mock_request.get(url_link_sample, status_code=400)
+        search_string = _build_search_string()
+        mock_request.get(f'{url_get_collection_instrument}?{search_string}', json=self.collection_instruments,
+                         complete_qs=True)
+
+        # When
+        response = self.app.get(f'/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}')
+
+        # Then
+        self.assertEqual(response.status_code, 400)
+
+    @requests_mock.mock()
+    def test_get_sample_returns_error(self, mock_request):
+        # Given
+        mock_request.get(url_get_survey_by_short_name, json=self.survey)
+        mock_request.get(url_ces, json=self.collection_exercises)
+        mock_request.get(url_ce, json=collection_exercise)
+        mock_request.get(url_link_sample, json=[sample_summary_id])
+        mock_request.get(url_sample_summary, status_code=400)
+        search_string = _build_search_string()
+        mock_request.get(f'{url_get_collection_instrument}?{search_string}', json=self.collection_instruments,
+                         complete_qs=True)
+
+        # When
+        response = self.app.get(f'/backstage-api/v1/collection-exercise/{test_short_name}/{test_period}')
 
         # Then
         self.assertEqual(response.status_code, 400)
